chore: add tab navigation to trial details page [DET-5070]
